Refactor backend api code
