+++ digiprod_gen/backend_api/api/common.py
@@ -1,0 +1,3 @@
+from digiprod_gen.backend_api.utils import initialise_config
+
+CONFIG = initialise_config("config/app-config.yaml")
