+++ digiprod_gen/backend_api/api/endpoints/browser.py
@@ -1,49 +1,39 @@
-import logging
 import time
-import io
-
 import sys
-from typing import List, Annotated
-from fastapi import FastAPI, Depends
-from fastapi.responses import StreamingResponse
+import logging
+from typing import List
 from functools import lru_cache
 
-from selenium.common.exceptions import NoSuchElementException, ElementNotInteractableException
-from digiprod_gen.backend_api.models.mba import MBAProduct, CrawlingMBARequest
+from fastapi import APIRouter
+from selenium.common import NoSuchElementException, ElementNotInteractableException
+
+from digiprod_gen.backend_api.api.common import  CONFIG
 from digiprod_gen.backend_api.browser.crawling import mba as mba_crawling
 from digiprod_gen.backend_api.browser.parser import mba as mba_parser
-from digiprod_gen.backend_api.browser.selenium_fns import SeleniumBrowser, wait_until_element_exists, get_full_page_screenshot
-from digiprod_gen.backend_api.utils.utils import delete_files_in_path, is_debug, initialise_config
+from digiprod_gen.backend_api.browser.selenium_fns import wait_until_element_exists, SeleniumBrowser
+from digiprod_gen.backend_api.models.mba import CrawlingMBARequest, MBAProduct
+from digiprod_gen.backend_api.utils import delete_files_in_path, is_debug
+
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 logger = logging.getLogger("BackendAPI")
 
-app = FastAPI()
-CONFIG = initialise_config("config/app-config.yaml")
+router = APIRouter()
 
 @lru_cache()
 def init_selenium_browser(session_id, proxy=None) -> SeleniumBrowser:
+    # New browser is started per session and per proxy
     logger.info(f"Init selenium browser with session_id {session_id}")
-    # TODO: Browser would be started with every api call. Better would be to start it per session user
     data_dir_path = CONFIG.browser.selenium_data_dir_path
     delete_files_in_path(data_dir_path)
     browser = SeleniumBrowser()
-    # session_state.get_marketplace_config().get_proxy_with_secrets(
-    #     st.secrets.proxy_perfect_privacy.user_name,
-    #     st.secrets.proxy_perfect_privacy.password)
     browser.setup(headless=not is_debug(),
                   data_dir_path=data_dir_path,
                   proxy=proxy
                   )
     return browser
 
-def init_selenium_browser_working() -> SeleniumBrowser:
-    # TODO: Browser would be started with every api call. Better would be to start it per session user
-    browser = SeleniumBrowser()
-    browser.setup()
-    return browser
-
-@app.post("/browser/crawling/mba_overview")
+@router.post("/crawling/mba_overview")
 async def crawl_mba_overview(request: CrawlingMBARequest, session_id: str) -> List[MBAProduct]:
     """ Searches mba overview page and change postcode in order to see correct products"""
     #, browser: Annotated[SeleniumBrowser, Depends(init_selenium_browser_working)]
@@ -78,7 +68,8 @@ async def crawl_mba_overview(request: CrawlingMBARequest, session_id: str) -> Li
     mba_products: List[MBAProduct] = mba_parser.extract_mba_products(browser.driver, request.marketplace)
     return mba_products
 
-@app.post("/browser/crawling/mba_product")
+
+@router.post("/crawling/mba_product")
 async def crawl_mba_product(mba_product: MBAProduct, session_id: str, proxy: str | None = None) -> MBAProduct:
     browser = init_selenium_browser(session_id, proxy)
     browser.driver.get(mba_product.product_url)
@@ -86,24 +77,8 @@ async def crawl_mba_product(mba_product: MBAProduct, session_id: str, proxy: str
     return mba_product
 
 
-@app.get("/status/browser_screenshot")
-async def get_browser_screenshot(session_id: str, proxy: str | None = None) -> StreamingResponse:
-    browser = init_selenium_browser(session_id, proxy)
-    screenshot_bytes = get_full_page_screenshot(browser.driver)
-    return StreamingResponse(io.BytesIO(screenshot_bytes), media_type="image/png")
-
-@app.get("/status/browser_settings")
-async def get_browser_settings(session_id: str, proxy: str | None = None) -> str:
-    browser = init_selenium_browser(session_id, proxy)
-    # Get the actual User-Agent used by the WebDriver
-    user_agent = browser.driver.execute_script("return navigator.userAgent")
-
-    # Access headers of a specific request (for example, the first request)
-    request = browser.driver.requests[0]
-    request_headers = request.headers
-    response_headers = request.response.headers
-
-    return f"""
-    User Agent: {user_agent}\n\nProxy Config: {browser.driver.proxy}\n\nRequest Headers: {request_headers}\n\nResponse Headers: {response_headers}
-    """
-
+def init_selenium_browser_working() -> SeleniumBrowser:
+    # TODO: Browser would be started with every api call. Better would be to start it per session user
+    browser = SeleniumBrowser()
+    browser.setup()
+    return browser
