+++ digiprod_gen/frontend/backend_caller.py
@@ -3,7 +3,7 @@ from typing import Any
 
 from fastapi.testclient import TestClient
 from digiprod_gen.backend.data_classes.config import BackendConfig
-from digiprod_gen.backend_api.api import app
+from digiprod_gen.backend_api.api.api import app
 
 class BackendCaller:
     """Gateway between frontend and backend"""
