+++ pyproject.toml
@@ -30,6 +30,7 @@ selenium-wire = "^5.1.0"
 [tool.poetry.group.backend.dependencies]
 fastapi = "0.103.2"
 httpx = "^0.25.0"
+uvicorn = "^0.23.2"
 
 [build-system]
 requires = ["poetry-core"]
